TIH/requirements.txt
asgiref==3.7.2
cffi==1.16.0
cryptography==41.0.7
distlib==0.3.7
Django==4.1.13
django-rest-knox==4.2.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.4.2
filelock==3.13.1
platformdirs==3.11.0
pycparser==2.21
pymongo==3.12.3
pytz==2023.3.post1
sqlparse==0.2.4
tzdata==2023.3
virtualenv==20.24.6
virtualenvwrapper-win==1.2.7
